+++ lib/src/db/lib_prod_archives.dart
@@ -2,6 +2,7 @@ library lib.src.db.lib_prod_archives;
 
 import 'dart:io';
 
+import 'package:dotenv/dotenv.dart' as dotenv;
 import 'package:elec_server/db_isone.dart';
 import 'package:elec_server/db_nyiso.dart';
 import 'package:elec_server/src/db/cme/cme_energy_settlements.dart';
@@ -130,6 +131,25 @@ IsoneBtmSolarArchive getIsoneHistoricalBtmSolarArchive() {
   }
   return IsoneBtmSolarArchive(dbConfig: dbConfig, dir: dir.path);
 }
+
+MonthlyAssetNcpcArchive getIsoneMonthlyAssetNcpcArchive() {
+  var config = ComponentConfig(
+      host: dotenv.env['MONGO_CONNECTION']!,
+      dbName: 'isoexpress',
+      collectionName: 'monthly_asset_ncpc');
+  final dir = '${Platform.environment['HOME'] ?? ''}/Downloads/Archive'
+      '/IsoExpress/GridReports/MonthlyAssetNcpc';
+  if (!Directory(dir).existsSync()) {
+    Directory(dir).createSync(recursive: true);
+  }
+
+  /// data starts on 4/1/2019
+  return MonthlyAssetNcpcArchive()
+    ..dbConfig = config
+    ..dir = dir
+    ..reportName = 'Monthly NCPC credits by Asset Report';
+}
+
 
 MraCapacityBidOfferArchive getIsoneMraBidOfferArchive() {
   final dir = '${Platform.environment['HOME'] ?? ''}/Downloads/Archive'